split cli.rb
